Support verification of encrypted images
An image's hash cannot be verified while the image is encrypted.  To
verify the hash of such an image, the image must be decrypted first
(without clearing the encrypted flag in the header).  This complicates
the API, as the caller now needs to pass in a set of encryption keys.

The fix is to split the image.Verify() function into several pieces:
    * VerifyStructure()
    * VerifyHash()
    * VerifySigs()
    * VerifyManifest()