mfg: Verify embedded images separately
Prior to this commit, the `"mfg".Mfg.VerifyManifest()` function
attempted to verify all of an mfg's embedded images.  This resulted in a
clunky and unfriendly API:

* The caller was required to pass in a set of public signing keys and
private encryption keys.
* There was no way to verify the manifest without also checking the
embedded images.  E.g., if the caller doesn't have access to encryption
keys, this function would fail even if the mfgimage is correct.

Now `VerifyManifest()` does not verify the embedded images.  To verify
the embedded images, the caller should call `"mfg".Mfg.ExtractImages()`
and call the image verification functions on each image separately.