New mfg manifest fields: target.extra, raw.extra
Add a new field to the `target` and `raw` mfg manifest items:
`extra_manifest`.  This field is a generic map that gets passed through
from the mfg definition.  This field is useful if the user needs to
convey additional information to some system that knows how to parse a
manufacturing manifest file.