image: Add support for section TLV
This patch adds support for the section TLV.  The Section TLV is a protected
TLV that supplies information about a specific elf section residing inside the
image.  There are some cases where only a image file is available and the
section information is required to do some post processing.  This TLV adds the
ability to deal with post processing in the absence of elf files.

Signed-off-by: Andy Gross <andy.gross@juul.com>